+++ RockWeb/Blocks/Examples/SampleData.ascx.cs
@@ -134,6 +134,11 @@ namespace RockWeb.Blocks.Examples
         private Dictionary<Guid, int> _peopleDictionary = new Dictionary<Guid, int>();
 
         /// <summary>
+        /// Holds a cached copy of the attribute Id for each person Guid
+        /// </summary>
+        private Dictionary<Guid, int> _peopleAliasDictionary = new Dictionary<Guid, int>();
+
+        /// <summary>
         /// Holds a cache of the person object
         /// </summary>
         private Dictionary<Guid, Person> _personCache = new Dictionary<Guid, Person>();
@@ -224,7 +229,7 @@ namespace RockWeb.Blocks.Examples
                     nbMessage.Text = string.Format(
 @"<p>Happy tire-kicking! The data is in your database. Hint: try <a href='{0}'>searching for the Decker family</a>.</p>
 <p>Here are some of the things you'll find in the sample data:</p>{1}",
-                        ResolveRockUrl( "~/Person/Search/name/Decker" ),
+                        ResolveRockUrl( "~/Person/Search/name/?SearchTerm=Decker" ),
                         GetStories( saveFile ) );
                     pnlInputForm.Visible = false;
                     RecordSuccess();
@@ -706,6 +711,7 @@ namespace RockWeb.Blocks.Examples
             var allFamilies = rockContext.Groups;
 
             List<Group> allGroups = new List<Group>();
+            var attendanceData = new Dictionary<Guid, List<Attendance>>();
 
             // Next create the family along with its members.
             foreach ( var elemFamily in elemFamilies.Elements( "family" ) )
@@ -726,7 +732,7 @@ namespace RockWeb.Blocks.Examples
                 // add their attendance data
                 if ( fabricateAttendance )
                 {
-                    AddFamilyAttendance( family, elemFamily, rockContext );
+                    AddFamilyAttendance( family, elemFamily, rockContext, attendanceData );
                 }
 
                 allGroups.Add( family );
@@ -754,7 +760,7 @@ namespace RockWeb.Blocks.Examples
                 {
                     foreach ( var attributeCache in gm.Person.Attributes.Select( a => a.Value ) )
                     {
-                        var newValue = gm.Person.AttributeValues[attributeCache.Key].FirstOrDefault();
+                        var newValue = gm.Person.AttributeValues[attributeCache.Key];
                         if ( newValue != null )
                         {
                             newValue.EntityId = gm.Person.Id;
@@ -777,9 +783,40 @@ namespace RockWeb.Blocks.Examples
             {
                 person.Aliases.Add( new PersonAlias { AliasPersonId = person.Id, AliasPersonGuid = person.Guid } );
             }
+            rockContext.ChangeTracker.DetectChanges();
+            rockContext.SaveChanges( disablePrePostProcessing: true );
+
+            // Put the person alias ids into the people alias dictionary for later use.
+            PersonAliasService personAliasService = new PersonAliasService( rockContext );
+            foreach ( var personAlias in personAliasService.Queryable( "Person" )
+                .Where( a =>
+                    _peopleDictionary.Keys.Contains( a.Person.Guid ) &&
+                    a.PersonId == a.AliasPersonId ) )
+            { 
+                _peopleAliasDictionary.Add( personAlias.Person.Guid, personAlias.Id );
+            }
 
             _stopwatch.Stop();
             _sb.AppendFormat( "{0:00}:{1:00}.{2:00} added person aliases<br/>", _stopwatch.Elapsed.Minutes, _stopwatch.Elapsed.Seconds, _stopwatch.Elapsed.Milliseconds / 10 );
+            _stopwatch.Start();
+
+            // Now that person aliases have been saved, save the attendance records
+            var attendanceService = new AttendanceService( rockContext );
+            var attendanceGuids = attendanceData.Select( a => a.Key ).ToList();
+            foreach ( var aliasKeyValue in _peopleAliasDictionary
+                .Where( a => attendanceGuids.Contains( a.Key )) )
+            {
+                foreach ( var attendance in attendanceData[aliasKeyValue.Key] )
+                {
+                    attendance.PersonAliasId = aliasKeyValue.Value;
+                    attendanceService.Add( attendance );
+                }
+            }
+            rockContext.ChangeTracker.DetectChanges();
+            rockContext.SaveChanges( disablePrePostProcessing: true );
+
+            _stopwatch.Stop();
+            _sb.AppendFormat( "{0:00}:{1:00}.{2:00} added attendance records<br/>", _stopwatch.Elapsed.Minutes, _stopwatch.Elapsed.Seconds, _stopwatch.Elapsed.Milliseconds / 10 );
             _stopwatch.Start();
         }
 
@@ -860,7 +897,7 @@ namespace RockWeb.Blocks.Examples
                     // Set the group location's GroupMemberPersonId if given (required?)
                     if ( elemGroup.Attribute( "meetsAtHomeOfPerson" ) != null )
                     {
-                        groupLocation.GroupMemberPersonId = _peopleDictionary[elemGroup.Attribute( "meetsAtHomeOfPerson" ).Value.AsGuid()];
+                        groupLocation.GroupMemberPersonAliasId = _peopleAliasDictionary[elemGroup.Attribute( "meetsAtHomeOfPerson" ).Value.AsGuid()];
                     }
 
                     group.GroupLocations.Add( groupLocation );
@@ -896,7 +933,37 @@ namespace RockWeb.Blocks.Examples
                 // Now we have to save changes in order for the attributes to be saved correctly.
                 rockContext.SaveChanges();
                 group.SaveAttributeValues( rockContext );
-            }
+
+                // Now add any group location schedules
+                LocationService locationService = new LocationService( rockContext );
+                ScheduleService scheduleService = new ScheduleService( rockContext );
+                Guid locationTypeMeetingLocationGuid = new Guid( Rock.SystemGuid.DefinedValue.GROUP_LOCATION_TYPE_MEETING_LOCATION );
+                var locationTypeMeetingLocationId = Rock.Web.Cache.DefinedValueCache.Read( locationTypeMeetingLocationGuid ).Id;
+
+                foreach ( var elemLocation in elemGroup.Elements( "location" ) )
+                {
+                    Guid locationGuid = elemLocation.Attribute( "guid" ).Value.Trim().AsGuid();
+                    Location location = locationService.Get( locationGuid );
+                    GroupLocation groupLocation = new GroupLocation();
+                    groupLocation.Location = location;
+                    groupLocation.GroupLocationTypeValueId = locationTypeMeetingLocationId;
+                    group.GroupLocations.Add( groupLocation );
+
+                    foreach ( var elemSchedule in elemLocation.Elements( "schedule" ) )
+                    {
+                        try
+                        {
+                            Guid scheduleGuid = elemSchedule.Attribute( "guid" ).Value.Trim().AsGuid();
+                            Schedule schedule = scheduleService.Get( scheduleGuid );
+                            groupLocation.Schedules.Add( schedule );
+                        }
+                        catch
+                        { }
+                    }
+                    TimeSpan ts = _stopwatch.Elapsed;
+                    _sb.AppendFormat( "{0:00}:{1:00}.{2:00} group location schedules added<br/>", ts.Minutes, ts.Seconds, ts.Milliseconds / 10 );
+                }
+             }
         }
 
         /// <summary>
@@ -963,6 +1030,8 @@ namespace RockWeb.Blocks.Examples
             PersonAliasService personAliasService = new PersonAliasService( rockContext );
             NoteService noteService = new NoteService( rockContext );
             AuthService authService = new AuthService( rockContext );
+            CommunicationService communicationService = new CommunicationService( rockContext );
+            CommunicationRecipientService communicationRecipientService = new CommunicationRecipientService( rockContext );
 
             foreach ( var elemFamily in families.Elements( "family" ) )
             {
@@ -992,6 +1061,18 @@ namespace RockWeb.Blocks.Examples
                             {
                                 phoneNumberService.Delete( phone );
                             }
+                        }
+
+                        // delete communication recipient
+                        foreach ( var recipient in communicationRecipientService.Queryable().Where( r => r.PersonAlias.PersonId == person.Id ) )
+                        {
+                            communicationRecipientService.Delete( recipient );
+                        }
+
+                        // delete communication
+                        foreach ( var communication in communicationService.Queryable().Where( c => c.SenderPersonAliasId == person.PrimaryAlias.Id ) )
+                        {
+                            communicationService.Delete( communication );
                         }
 
                         // delete person viewed records
@@ -1097,15 +1178,12 @@ namespace RockWeb.Blocks.Examples
             if ( group.AttributeValues != null )
             {
                 var attributeValueService = new AttributeValueService( rockContext );
-                foreach ( KeyValuePair<string, List<AttributeValue>> entry in group.AttributeValues )
+                foreach ( var entry in group.AttributeValues )
                 {
-                    foreach ( AttributeValue value in entry.Value )
+                    var attributeValue = attributeValueService.GetByAttributeIdAndEntityId( entry.Value.AttributeId, group.Id );
+                    if ( attributeValue != null )
                     {
-                        var attributeValues = attributeValueService.GetByAttributeIdAndEntityId( value.AttributeId, group.Id ).ToList();
-                        foreach ( var attributeValue in attributeValues )
-                        {
-                            attributeValueService.Delete( attributeValue );
-                        }
+                        attributeValueService.Delete( attributeValue );
                     }
                 }
             }
@@ -1152,7 +1230,8 @@ namespace RockWeb.Blocks.Examples
         /// <param name="family">The family.</param>
         /// <param name="elemFamily">The elem family.</param>
         /// <param name="rockContext">The rock context.</param>
-        private void AddFamilyAttendance( Group family, XElement elemFamily, RockContext rockContext )
+        /// <param name="attendanceData">The attendance data.</param>
+        private void AddFamilyAttendance( Group family, XElement elemFamily, RockContext rockContext, Dictionary<Guid, List<Attendance>> attendanceData )
         {
             // return from here if there's not startingAttendance date
             if ( elemFamily.Attribute( "startingAttendance" ) == null )
@@ -1223,7 +1302,7 @@ namespace RockWeb.Blocks.Examples
                 }
             }
 
-            CreateAttendance( family.Members, startingDate, endDate, pctAttendance, pctAttendedRegularService, scheduleId, altScheduleId );
+            CreateAttendance( family.Members, startingDate, endDate, pctAttendance, pctAttendedRegularService, scheduleId, altScheduleId, attendanceData );
         }
 
         /// <summary>
@@ -1233,14 +1312,16 @@ namespace RockWeb.Blocks.Examples
         /// between the scheduleId and altScheduleId based on the percentage (pctAttendedRegularService)
         /// given.
         /// </summary>
-        /// <param name="familyMembers"></param>
+        /// <param name="familyMembers">The family members.</param>
         /// <param name="startingDate">The first date of attendance</param>
         /// <param name="endDate">The end date of attendance</param>
-        /// <param name="pctAttendance"></param>
-        /// <param name="pctAttendedRegularService"></param>
-        /// <param name="scheduleId"></param>
-        /// <param name="altScheduleId"></param>
-        private void CreateAttendance( ICollection<GroupMember> familyMembers, DateTime startingDate, DateTime endDate, int pctAttendance, int pctAttendedRegularService, int scheduleId, int altScheduleId )
+        /// <param name="pctAttendance">The PCT attendance.</param>
+        /// <param name="pctAttendedRegularService">The PCT attended regular service.</param>
+        /// <param name="scheduleId">The schedule identifier.</param>
+        /// <param name="altScheduleId">The alt schedule identifier.</param>
+        /// <param name="attendanceData">The attendance data.</param>
+        private void CreateAttendance( ICollection<GroupMember> familyMembers, DateTime startingDate, DateTime endDate, 
+            int pctAttendance, int pctAttendedRegularService, int scheduleId, int altScheduleId, Dictionary<Guid, List<Attendance>> attendanceData )
         {
             // foreach weekend between the starting and ending date...
             for ( DateTime date = startingDate; date <= endDate; date = date.AddDays( 7 ) )
@@ -1295,7 +1376,6 @@ namespace RockWeb.Blocks.Examples
                         GroupId = item.GroupId,
                         LocationId = item.LocationId,
                         DeviceId = _kioskDeviceId,
-                        PersonId = member.PersonId,
                         AttendanceCode = attendanceCode,
                         StartDateTime = checkinDateTime,
                         EndDateTime = null,
@@ -1303,7 +1383,11 @@ namespace RockWeb.Blocks.Examples
                         CampusId = 1
                     };
 
-                    member.Person.Attendances.Add( attendance );
+                    if ( !attendanceData.Keys.Contains( member.Person.Guid ))
+                    {
+                        attendanceData.Add( member.Person.Guid, new List<Attendance>());
+                    }
+                    attendanceData[member.Person.Guid].Add( attendance);
                 }
             }
         }
